Tempo de Entrega de Refrigerantes

Exemplo: Uma engarrafadora de refrigerantes está analisando as rotas de serviço das máquinas de venda automática em seu sistema de distribuição. A empresa está interessada em prever a quantidade de tempo necessária para a rota motorista para atender as máquinas de venda automática em um ponto de venda. Esta atividade de serviço inclui estocar a máquina com produtos de bebidas e a realização de pequenas manutenções. O engenheiro industrial responsável pelo estudo sugeriu que duas variáveis mais importantes que afetam o tempo de entrega são o número de caixas de produto estocado e a distância percorrida pelo roteirista . Foram coletadas 25 observações sobre o tempo de entrega.

Format

Data frame com 25 linhas e 3 variáveis:

tempo: tempo de entrega (em minutos)

caixas: número de caixas a serem entregues

distancia: distância percorrida (em pés)

References

Montgomery, D. C., Peck, E. A., and Vining, G. G. (2012) Introduction to Linear Regression Analysis, 5th Edition, Boca Raton: Wiley.

Author

Fábio N. Demarqui fndemarqui@est.ufmg.br